minesweepertest -> minesweeper로 변경
