--- src/main/kotlin/minesweeper/MineSweeperApplication.kt
@@ -1,7 +1,0 @@
-package minesweeper
-
-import minesweeper.controller.MineSweeperController
-
-fun main() {
-    MineSweeperController().start()
-}
