--- src/main/kotlin/minesweeper/domain/Block.kt
@@ -1,3 +1,0 @@
-package minesweeper.domain
-
-sealed class Block
