+++ src/main/kotlin/minesweeperpractice/MineSweeperApplication.kt
@@ -1,0 +1,7 @@
+package minesweeperpractice
+
+import minesweeperpractice.controller.MineSweeperController
+
+fun main() {
+    MineSweeperController().start()
+}
