+++ src/main/kotlin/minesweeperpractice/domain/Block.kt
@@ -1,0 +1,3 @@
+package minesweeperpractice.domain
+
+sealed class Block
