+++ src/main/kotlin/minesweeperpractice/view/model/BlockModel.kt
@@ -1,4 +1,4 @@
-package minesweeper.view.model
+package minesweeperpractice.view.model
 
 sealed class BlockModel {
     abstract override fun toString(): String
