+++ src/test/kotlin/minesweeper/GameMapTest.kt
@@ -1,4 +1,4 @@
-package minesweepertest
+package minesweeper
 
 import io.kotest.core.spec.style.ExpectSpec
