+++ src/test/kotlin/minesweeperpractice/domain/GameMapTest.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeperpractice.domain
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.ExpectSpec
